New run flow

+++ profit/pre.py
@@ -22,21 +22,26 @@ def fill_run_dir(eval_points, template_dir='template/', run_dir='run/', param_fi
 
         # .zfill(3) is an option that forces krun to have 3 digits
         run_dir_single = os.path.join(run_dir, str(krun).zfill(3))
-        if os.path.exists(run_dir_single):
-            if overwrite:
-                rmtree(run_dir_single)
-            else:
-                raise RuntimeError('Run directory not empty: {}'.format(run_dir_single))
-        copy_template(template_dir, run_dir_single)
+        fill_run_dir_single(eval_points[krun], template_dir, run_dir_single, param_files, overwrite)
 
-        fill_template(run_dir_single, eval_points[krun], param_files=param_files)
+
+def fill_run_dir_single(params, template_dir, run_dir_single, param_files=None, overwrite=False,
+                        ignore_path_exists=False):
+    if os.path.exists(run_dir_single) and not ignore_path_exists:  # ToDo: make ignore_path_exists default
+        if overwrite:
+            rmtree(run_dir_single)
+        else:
+            raise RuntimeError('Run directory not empty: {}'.format(run_dir_single))
+    copy_template(template_dir, run_dir_single)
+
+    fill_template(run_dir_single, params, param_files=param_files)
 
 
 def copy_template(template_dir, out_dir, dont_copy=None):
     """ TODO: explain dont_copy patterns """
 
     if dont_copy:
-        copytree(template_dir, out_dir, ignore=ignore_patterns(*dont_copy))
+        copytree(template_dir, out_dir, symlinks=True, ignore=ignore_patterns(*dont_copy))
     else:
         copytree(template_dir, out_dir, symlinks=True)
     convert_relative_symlinks(template_dir, out_dir)
@@ -62,16 +67,22 @@ def fill_template(out_dir, params, param_files=None):
         for filename in files:
             if not param_files or filename in param_files:
                 filepath = os.path.join(root, filename)
-                with open(filepath, 'r') as f:
-                    content = f.read()
-                    # Escape '{*}' for e.g. json templates by replacing it with '{{*}}'.
-                    # Variables then have to be declared as '{{*}}' which is replaced by a single '{*}'.
-                    if '{{' in content:
-                        content = content.replace('{{', '§').replace('}}', '§§')\
-                            .replace('{', '{{').replace('}', '}}').replace('§§', '}').replace('§', '{')
-                    content = content.format_map(util.SafeDict(rec2dict(params)))
-                with open(filepath, 'w') as f:
-                    f.write(content)
+                fill_template_file(filepath, filepath, params)
+
+
+def fill_template_file(template_filepath, output_filepath, params):
+    with open(template_filepath, 'r') as f:
+        content = f.read()
+        # Escape '{*}' for e.g. json templates by replacing it with '{{*}}'.
+        # Variables then have to be declared as '{{*}}' which is replaced by a single '{*}'.
+        pre, post = '{', '}'
+        if '{{' in content:
+            content = content.replace('{{', '§').replace('}}', '§§') \
+                .replace('{', '{{').replace('}', '}}').replace('§§', '}').replace('§', '{')
+            pre, post = '{{', '}}'
+        content = content.format_map(util.SafeDict.from_params(params, pre=pre, post=post))
+    with open(output_filepath, 'w') as f:
+        f.write(content)
 
 
 def get_eval_points(config):
